wrapper around PyObject
