+++ samples/dotnet/PythonSkillRunner/pythonskills/randomActivitySkill.py
@@ -5,4 +5,4 @@ class RandomActivitySkill:
     def getRandomActivity():
         result = requests.get('https://www.boredapi.com/api/activity').text;
         activityObject = json.loads(result);
-        return activityObject
+        return activityObject['activity']
